requirements.txt
Appium-Python-Client==1.3.0
mysql-connector-python==8.0.31
pytest==7.1.3
# selenium==3.141.0 # For Desktop Electron Mobile
selenium==4.5.0 #For Web ONLY!!!
requests==2.25.1
eyes-selenium==5.12.3